refactor: modularize scripts.js and improve validation
