+++ constants.js
@@ -1,0 +1,14 @@
+export const FOLDER_VARIANTS = {
+  "Data-Base": ["Data-Base", "Data Base", "DataBase", "data-base"],
+  "C#": ["C%23", "C#", "c%23", "c-sharp", "CSharp"],
+  QA: ["QA", "qa", "Qa", "Q&A"],
+};
+
+export const REVERSE_FOLDER_MAPPING = {
+  "c%23": "C#",
+  "c-sharp": "C#",
+  csharp: "C#",
+  "data base": "Data-Base",
+  database: "Data-Base",
+  "q&a": "QA",
+};
