+++ index.html
@@ -35,6 +35,6 @@
       </main>
     </div>
 
-    <script src="scripts.js"></script>
+    <script type="module" src="main.js"></script>
   </body>
 </html>
